TIP-863: move Group Transalation to Enrichment
