+++ src/Akeneo/Pim/Enrichment/Bundle/DependencyInjection/CompilerPass/ResolveDoctrineTargetModelPass.php
@@ -8,6 +8,7 @@ use Akeneo\Pim\Enrichment\Component\Category\Model\CategoryInterface;
 use Akeneo\Pim\Enrichment\Component\Category\Model\CategoryTranslationInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\CompletenessInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\GroupInterface;
+use Akeneo\Pim\Enrichment\Component\Product\Model\GroupTranslationInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductAssociationInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Pim\Enrichment\Component\Product\Model\ProductModelAssociationInterface;
@@ -34,6 +35,7 @@ class ResolveDoctrineTargetModelPass extends AbstractResolveDoctrineTargetModelP
             CategoryTranslationInterface::class => 'pim_catalog.entity.category_translation.class',
             CompletenessInterface::class => 'pim_catalog.entity.completeness.class',
             GroupInterface::class => 'pim_catalog.entity.group.class',
+            GroupTranslationInterface::class => 'pim_catalog.entity.group_translation.class',
             ProductInterface::class => 'pim_catalog.entity.product.class',
             ProductAssociationInterface::class => 'pim_catalog.entity.association.class',
             ProductModelInterface::class => 'pim_catalog.entity.product_model.class',
