+++ src/Akeneo/Pim/Enrichment/Component/Product/Model/GroupInterface.php
@@ -2,11 +2,9 @@
 
 namespace Akeneo\Pim\Enrichment\Component\Product\Model;
 
-use Akeneo\Pim\Enrichment\Component\Product\Model\ProductInterface;
 use Akeneo\Tool\Component\Localization\Model\TranslatableInterface;
 use Akeneo\Tool\Component\StorageUtils\Model\ReferableInterface;
 use Akeneo\Tool\Component\Versioning\Model\VersionableInterface;
-use Doctrine\Common\Collections\ArrayCollection;
 use Pim\Component\Catalog\Model\GroupTypeInterface;
 use Symfony\Component\Validator\GroupSequenceProviderInterface;
