+++ src/Akeneo/Pim/Enrichment/Component/Product/Model/GroupTranslation.php
@@ -1,9 +1,8 @@
 <?php
 
-namespace Pim\Bundle\CatalogBundle\Entity;
+namespace Akeneo\Pim\Enrichment\Component\Product\Model;
 
 use Akeneo\Tool\Component\Localization\Model\AbstractTranslation;
-use Pim\Component\Catalog\Model\GroupTranslationInterface;
 
 /**
  * Group translation
