+++ src/Pim/Bundle/CatalogBundle/Resources/config/entities.yml
@@ -11,7 +11,7 @@ parameters:
     pim_catalog.entity.category.class:                     Akeneo\Pim\Enrichment\Component\Category\Model\Category
     pim_catalog.entity.category_translation.class:         Akeneo\Pim\Enrichment\Component\Category\Model\CategoryTranslation
     pim_catalog.entity.group.class:                        Akeneo\Pim\Enrichment\Component\Product\Model\Group
-    pim_catalog.entity.group_translation.class:            Pim\Bundle\CatalogBundle\Entity\GroupTranslation
+    pim_catalog.entity.group_translation.class:            Akeneo\Pim\Enrichment\Component\Product\Model\GroupTranslation
     pim_catalog.entity.group_type.class:                   Pim\Bundle\CatalogBundle\Entity\GroupType
     pim_catalog.entity.group_type_translation.class:       Pim\Bundle\CatalogBundle\Entity\GroupTypeTranslation
     pim_catalog.entity.attribute.class:                    Akeneo\Pim\Structure\Component\Model\Attribute
