+++ src/Pim/Bundle/CatalogBundle/Resources/config/models.yml
@@ -23,5 +23,5 @@ parameters:
     pim_catalog.model.association_type_translation.interface: Pim\Component\Catalog\Model\AssociationTypeTranslationInterface
     pim_catalog.model.attribute_group_translation.interface:  Akeneo\Pim\Structure\Component\Model\AttributeGroupTranslationInterface
     pim_catalog.model.attribute_translation.interface:        Akeneo\Pim\Structure\Component\Model\AttributeTranslationInterface
-    pim_catalog.model.group_translation.interface:            Pim\Component\Catalog\Model\GroupTranslationInterface
+    pim_catalog.model.group_translation.interface:            Akeneo\Pim\Enrichment\Component\Product\Model\GroupTranslationInterface
     pim_catalog.model.group_type_translation.interface:       Pim\Component\Catalog\Model\GroupTypeTranslationInterface
